Implement editor panels, transform gizmo, easing library, and rounded quad SDF
Introduces a full in-engine editor layer on top of the Vulkan renderer.

Editor / Panel system (editor.c, editor.h)
- Four animated slide-in panels: Files (bottom), Inspector (right),
  Console (top), Hierarchy (left)
- Per-panel EaseFn, exponential approach animation (frame-rate
  independent), min/max size constraints
- Panels toggled via Alt+hjkl keybindings
- Inspector: displays transform, material (PBR factors, IOR,
  transmission), geometry, and skeletal/morph animation data for the
  selected mesh
- Hierarchy: mesh list with alpha-mode dot indicators and selection
  highlight
- File manager: directory listing with tile grid layout, UTF-8 ellipsis
  truncation, size labels, navigate-up support
- MapleMono NF 18px as the editor font

Transform gizmo (gizmo.c, gizmo.h)
- Translate (arrows + plane squares), Rotate (camera-facing arcs),
  Scale (cube tips) modes; cycle with G
- Screen-space hit testing: axis segments, plane square centers, arc
  polyline sampling
- Drag applies in world space via virtual constraint planes (axis) or
  fixed planes (XY/XZ/YZ); rotation and scale pivot at mesh origin
- Mesh picking via GPU depth buffer readback (renderer_read_depth_at)
  unprojected to world space, AABB containment test with volume
  tie-breaking for overlapping meshes
- Gizmo rendered at fixed near-plane distance to avoid depth occlusion,
  scale compensated to preserve screen footprint

Easing library (easing.c, easing.h)
- Full Robert Penner family (sine, quad, cubic, quart, quint, expo,
  circ, back, elastic, bounce) plus smooth/smoother/smoothest step,
  spring, arch, punch, stepped, zigzag, mirror, chain
- CSS cubic-bezier solver via 4-iteration Newton-Raphson
- Parametric back (custom s), elastic (amplitude + period), spring
  (stiffness + damping ratio)
- SSE2 / AVX2 / NEON batch paths for linear, quad, cubic, smooth_step
- Float and double dispatch tables initialised by ease_init()
- ease_remap, ease_mirror, ease_chain higher-order helpers

Renderer: rounded quad primitive (renderer.c, renderer.h, shaders/)
- New Vertex2D fields: size, cornerRadius (TL/TR/BR/BL), borderThickness,
  borderColor; Vulkan vertex input updated to 8 attributes
- exQuad2D(): emits textureIndex=-2 quads; frag shader computes
  per-quadrant SDF with anti-aliased smoothstep and inner/outer border
- quad2D and texture2D updated to zero-initialise new fields

PBR shader cleanup (shaders/pbr.frag)
- Extracted tonemapACES / inverseACES as named functions
- shadowVisibility replaces ShadowCalculation; cascade bias via bit-shift
- iblAmbient: explicit Y-flip for cubemap samples, cleaner specOcclusion
- getTransmittedColor: safe near-plane clamp, view-transpose instead of
  inverse, frosted glass weight corrected (0.1667)
- Dispersion: glTF-spec half-delta split rather than Abbe number
- Transmission block: diffuse/transmission mix in linear space, single
  tonemapACES call for both paths

Misc
- New "obsidian" theme with x/y/z axis colours and semantic success/
  warning/error palette; theme struct gains bg_alt, bg_deep, text_dim,
  border, accent, axis colour fields
- GLTF compat list updated: Fox, RiggedFigure, RecursiveSkeletons marked
  PASS; several previously FAILing assets promoted
- vertico.c: fix signed/unsigned comparison in selected_index check

+++ vertico.c
@@ -544,7 +544,7 @@ void vertico_render() {
         int idx = vertico.scroll_offset + i;
         if (idx >= (int)vertico.filtered_count) break;
 
-        bool is_selected = (idx == vertico.selected_index);
+        bool is_selected = ((int)idx == vertico.selected_index);
 
         // Selection highlight
         if (is_selected) {
